Added audio recorder support for tab key.
Tab key now works if embedded within a page that contains appropriate JavaScript. The page must insert tabbable elements before and after the applet and call an applet method if the user tabs to those elements so the applet can take control; then, the page must include a function which the applet can call when it wants to move the focus outside of itself.

If you don't include the 'focus hack ID' parameter to the applet, then it works as before (no tabbing support).

The embedded recorder already had tab support; this is an attempt to generalise that support, although I also had to extend it because there were extra problems.

(This is required because browsers/Oracle do not as standard provide any support for easily tabbing between browser and applets.)

OU bug 10408.